+++ kedro_mlflow/framework/cli/cli.py
@@ -1,16 +1,20 @@
-import os
 import subprocess
 from pathlib import Path
 
 import click
-from kedro import __file__ as KEDRO_PATH
+from kedro.framework.context import load_context
 
-from kedro_mlflow.framework.cli.cli_utils import (
-    render_jinja_template,
-    write_jinja_template,
-)
+from kedro_mlflow.framework.cli.cli_utils import write_jinja_template
 from kedro_mlflow.framework.context import get_mlflow_config
-from kedro_mlflow.utils import _already_updated, _get_project_globals, _is_kedro_project
+from kedro_mlflow.utils import _already_updated, _is_kedro_project
+
+try:
+    from kedro.framework.context import get_static_project_data
+except ImportError:  # pragma: no cover
+    from kedro_mlflow.utils import (
+        _get_project_globals as get_static_project_data,  # pragma: no cover
+    )
+
 
 TEMPLATE_FOLDER_PATH = Path(__file__).parent.parent.parent / "template" / "project"
 
@@ -45,15 +49,13 @@ class KedroClickGroup(click.Group):
 
 @click.group(name="Mlflow")
 def commands():
-    """Kedro plugin for interactions with mlflow.
-    """
+    """Kedro plugin for interactions with mlflow."""
     pass  # pragma: no cover
 
 
 @commands.command(name="mlflow", cls=KedroClickGroup)
 def mlflow_commands():
-    """Use mlflow-specific commands inside kedro project.
-    """
+    """Use mlflow-specific commands inside kedro project."""
     pass
 
 
@@ -87,7 +89,9 @@ def init(force, silent):
 
     # get constants
     project_path = Path().cwd()
-    project_globals = _get_project_globals()
+    project_globals = get_static_project_data(project_path)
+    context = load_context(project_path)
+    conf_root = context.CONF_ROOT
 
     # mlflow.yml is just a static file,
     # but the name of the experiment is set to be the same as the project
@@ -95,70 +99,13 @@ def init(force, silent):
     write_jinja_template(
         src=TEMPLATE_FOLDER_PATH / mlflow_yml,
         is_cookiecutter=False,
-        dst=project_path / "conf" / "base" / mlflow_yml,
-        python_package=project_globals["python_package"],
+        dst=project_path / conf_root / "base" / mlflow_yml,
+        python_package=project_globals["package_name"],
     )
     if not silent:
         click.secho(
-            click.style("'conf/base/mlflow.yml' successfully updated.", fg="green")
-        )
-    # make a check whether the project run.py is strictly identical to the template
-    # if yes, replace the script by the template silently
-    # if no, raise a warning and send a message to INSERT_DOC_URL
-    flag_erase_runpy = force
-    runpy_project_path = (
-        project_path
-        / "src"
-        / (Path(project_globals["context_path"]).parent.as_posix() + ".py")
-    )
-    if not force:
-        kedro_path = Path(KEDRO_PATH).parent
-        runpy_template_path = (
-            kedro_path
-            / "templates"
-            / "project"
-            / "{{ cookiecutter.repo_name }}"
-            / "src"
-            / "{{ cookiecutter.python_package }}"
-            / "run.py"
-        )
-        kedro_runpy_template = render_jinja_template(
-            src=runpy_template_path,
-            is_cookiecutter=True,
-            python_package=project_globals["python_package"],
-            project_name=project_globals["project_name"],
-            kedro_version=project_globals["kedro_version"],
-        )
-
-        with open(runpy_project_path, mode="r") as file_handler:
-            kedro_runpy_project = file_handler.read()
-
-        # beware : black formatting could change slightly this test which is very strict
-        if kedro_runpy_project == kedro_runpy_template:
-            flag_erase_runpy = True
-
-    if flag_erase_runpy:
-        os.remove(runpy_project_path)
-        write_jinja_template(
-            src=TEMPLATE_FOLDER_PATH / "run.py",
-            dst=runpy_project_path,
-            is_cookiecutter=True,
-            python_package=project_globals["python_package"],
-            project_name=project_globals["project_name"],
-            kedro_version=project_globals["kedro_version"],
-        )
-        if not silent:
-            click.secho(click.style("'run.py' successfully updated", fg="green"))
-    else:
-        click.secho(
             click.style(
-                "You have modified your 'run.py' since project creation.\n"
-                + "In order to use kedro-mlflow, you must either:\n"
-                + "    -  set up your run.py with the following instructions :\n"
-                + "INSERT_DOC_URL\n"
-                + "    - call the following command:\n"
-                + "$ kedro mlflow init --force",
-                fg="yellow",
+                f"'{conf_root}/base/mlflow.yml' successfully updated.", fg="green"
             )
         )
 
@@ -179,13 +126,16 @@ def init(force, silent):
 )
 def ui(project_path, env):
     """Opens the mlflow user interface with the
-        project-specific settings of mlflow.yml. This interface
-        enables to browse and compares runs.
+    project-specific settings of mlflow.yml. This interface
+    enables to browse and compares runs.
 
     """
 
+    if not project_path:
+        project_path = Path().cwd()
+    context = load_context(project_path=project_path, env=env)
     # the context must contains the self.mlflow attribues with mlflow configuration
-    mlflow_conf = get_mlflow_config(project_path=project_path, env=env)
+    mlflow_conf = get_mlflow_config(context)
 
     # call mlflow ui with specific options
     # TODO : add more options for ui
@@ -196,8 +146,7 @@ def ui(project_path, env):
 
 @mlflow_commands.command()
 def run():
-    """Re-run an old run with mlflow-logged info.
-    """
+    """Re-run an old run with mlflow-logged info."""
 
     # TODO (HARD) : define general assumptions to check whether a run
     #  is reproductible or not
@@ -211,13 +160,11 @@ def run():
 
 @mlflow_commands.command()
 def new():
-    """Create a new kedro project with updated template.
-    """
+    """Create a new kedro project with updated template."""
     raise NotImplementedError  # pragma: no cover
 
 
 class KedroMlflowCliError(Exception):
-    """ kedro-mlflow cli specific error
-    """
+    """kedro-mlflow cli specific error"""
 
     pass