+++ kedro_mlflow/pipeline/__init__.py
@@ -1,5 +1,6 @@
 from .pipeline_ml import (
     KedroMlflowPipelineMLDatasetsError,
     KedroMlflowPipelineMLInputsError,
+    KedroMlflowPipelineMLOutputsError,
 )
-from .pipeline_ml_factory import pipeline_ml, pipeline_ml_factory
+from .pipeline_ml_factory import pipeline_ml_factory