+++ kedro_mlflow/pipeline/pipeline_ml_factory.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 from typing import Any, Dict, Optional, Union
-from warnings import warn
 
 from kedro.pipeline import Pipeline
+from mlflow.models import ModelSignature
 
 from kedro_mlflow.pipeline.pipeline_ml import PipelineML
 
@@ -13,6 +13,7 @@ def pipeline_ml_factory(
     input_name: str = None,
     conda_env: Optional[Union[str, Path, Dict[str, Any]]] = None,
     model_name: Optional[str] = "model",
+    model_signature: Union[ModelSignature, str, None] = "auto",
 ) -> PipelineML:
     """This function is a helper to create `PipelineML`
     object directly from two Kedro `Pipelines` (one of
@@ -47,6 +48,13 @@ def pipeline_ml_factory(
         model_name (Union[str, None], optional): The name of
             the folder where the model will be stored in
             remote mlflow. Defaults to "model".
+        model_signature (Union[ModelSignature, bool]): The mlflow
+             signature of the input dataframe common to training
+             and inference.
+                   - If 'auto', it is infered automatically
+                   - If None, no signature is used
+                   - if a `ModelSignature` instance, passed
+                   to the underlying dataframe
 
     Returns:
         PipelineML: A `PipelineML` which is automatically
@@ -61,26 +69,6 @@ def pipeline_ml_factory(
         input_name=input_name,
         conda_env=conda_env,
         model_name=model_name,
+        model_signature=model_signature,
     )
     return pipeline
-
-
-def pipeline_ml(
-    training: Pipeline,
-    inference: Pipeline,
-    input_name: str = None,
-    conda_env: Optional[Union[str, Path, Dict[str, Any]]] = None,
-    model_name: Optional[str] = "model",
-) -> PipelineML:
-
-    deprecation_msg = (
-        "'pipeline_ml' is now deprecated and "
-        "has been renamed 'pipeline_ml_factory' "
-        "in 'kedro-mlflow>=0.3.0'. "
-        "\nPlease change your 'pipeline.py' or 'hooks.py' entries "
-        "accordingly, since 'pipeline_ml' will be removed in next release."
-    )
-
-    warn(deprecation_msg, DeprecationWarning, stacklevel=2)
-
-    return pipeline_ml_factory(training, inference, input_name, conda_env, model_name)