+++ tests/framework/hooks/test_pipeline_hook.py
@@ -1,13 +1,15 @@
 import sys
 
 import mlflow
+import pandas as pd
 import pytest
 import yaml
 from kedro.extras.datasets.pickle import PickleDataSet
-from kedro.framework.context import KedroContext
+from kedro.framework.context import KedroContext, load_context
 from kedro.io import DataCatalog, MemoryDataSet
 from kedro.pipeline import Pipeline, node
 from kedro.runner import SequentialRunner
+from mlflow.models import infer_signature
 from mlflow.tracking import MlflowClient
 
 from kedro_mlflow.framework.context import get_mlflow_config
@@ -16,7 +18,7 @@ from kedro_mlflow.framework.hooks.pipeline_hook import (
     _format_conda_env,
     _generate_kedro_command,
 )
-from kedro_mlflow.io import MlflowMetricsDataSet
+from kedro_mlflow.io.metrics import MlflowMetricsDataSet
 from kedro_mlflow.pipeline import pipeline_ml_factory
 from kedro_mlflow.pipeline.pipeline_ml import PipelineML
 
@@ -79,7 +81,9 @@ def env_from_environment(environment_path, env_from_dict):
     with open(environment_path, mode="w") as file_handler:
         yaml.dump(env_from_dict, file_handler)
 
-    return env_from_dict
+    env_from_environment = env_from_dict
+
+    return env_from_environment
 
 
 @pytest.mark.parametrize(
@@ -123,8 +127,8 @@ def dummy_pipeline():
     def train_fun(data, param):
         return 2
 
-    def metric_fun():
-        return {"metric": {"value": 1.1, "step": 0}}
+    def metric_fun(data, model):
+        return {"metric_key": {"value": 1.1, "step": 0}}
 
     def predict_fun(model, data):
         return data * model
@@ -143,8 +147,18 @@ def dummy_pipeline():
                 outputs="model",
                 tags=["training"],
             ),
-            node(func=metric_fun, inputs=None, outputs="metrics",),
-            node(func=metric_fun, inputs=None, outputs="another_metrics",),
+            node(
+                func=metric_fun,
+                inputs=["model", "data"],
+                outputs="my_metrics",
+                tags=["training"],
+            ),
+            node(
+                func=metric_fun,
+                inputs=["model", "data"],
+                outputs="another_metrics",
+                tags=["training"],
+            ),
             node(
                 func=predict_fun,
                 inputs=["model", "data"],
@@ -173,15 +187,22 @@ def dummy_pipeline_ml(dummy_pipeline, env_from_dict):
 def dummy_catalog(tmp_path):
     dummy_catalog = DataCatalog(
         {
-            "raw_data": MemoryDataSet(1),
+            "raw_data": MemoryDataSet(pd.DataFrame(data=[1], columns=["a"])),
             "params:unused_param": MemoryDataSet("blah"),
             "data": MemoryDataSet(),
             "model": PickleDataSet((tmp_path / "model.csv").as_posix()),
-            "metrics": MlflowMetricsDataSet(),
+            "my_metrics": MlflowMetricsDataSet(),
             "another_metrics": MlflowMetricsDataSet(prefix="foo"),
         }
     )
     return dummy_catalog
+
+
+@pytest.fixture
+def dummy_signature(dummy_catalog, dummy_pipeline_ml):
+    input_data = dummy_catalog.load(dummy_pipeline_ml.input_name)
+    dummy_signature = infer_signature(input_data)
+    return dummy_signature
 
 
 @pytest.fixture
@@ -258,19 +279,22 @@ def test_mlflow_pipeline_hook_with_different_pipeline_types(
     pipeline_hook.before_pipeline_run(
         run_params=dummy_run_params, pipeline=pipeline_to_run, catalog=dummy_catalog
     )
-    runner.run(pipeline_to_run, dummy_catalog, dummy_run_params["run_id"])
+    runner.run(pipeline_to_run, dummy_catalog)
     run_id = mlflow.active_run().info.run_id
     pipeline_hook.after_pipeline_run(
         run_params=dummy_run_params, pipeline=pipeline_to_run, catalog=dummy_catalog
     )
     # test : parameters should have been logged
-    mlflow_conf = get_mlflow_config(tmp_path)
+    context = load_context(tmp_path)
+    mlflow_conf = get_mlflow_config(context)
     mlflow_client = MlflowClient(mlflow_conf.mlflow_tracking_uri)
     run_data = mlflow_client.get_run(run_id).data
+
     # all run_params are recorded as tags
     for k, v in dummy_run_params.items():
         if v:
             assert run_data.tags[k] == str(v)
+
     # params are not recorded because we don't have MlflowNodeHook here
     # and the model should not be logged when it is not a PipelineML
     nb_artifacts = len(mlflow_client.list_artifacts(run_id))
@@ -278,10 +302,160 @@ def test_mlflow_pipeline_hook_with_different_pipeline_types(
         assert nb_artifacts == 1
     else:
         assert nb_artifacts == 0
+
     # Check if metrics datasets have prefix with its names.
     # for metric
-    assert dummy_catalog._data_sets["metrics"]._prefix == "metrics"
+    assert dummy_catalog._data_sets["my_metrics"]._prefix == "my_metrics"
     assert dummy_catalog._data_sets["another_metrics"]._prefix == "foo"
+
+    if isinstance(pipeline_to_run, PipelineML):
+        trained_model = mlflow.pyfunc.load_model(f"runs:/{run_id}/model")
+        assert trained_model.metadata.signature.to_dict() == {
+            "inputs": '[{"name": "a", "type": "long"}]',
+            "outputs": None,
+        }
+
+
+def test_mlflow_pipeline_hook_metrics_with_run_id(
+    mocker,
+    monkeypatch,
+    tmp_path,
+    config_dir,
+    env_from_dict,
+    dummy_pipeline_ml,
+    dummy_run_params,
+    dummy_mlflow_conf,
+):
+    # config_with_base_mlflow_conf is a conftest fixture
+    mocker.patch("kedro_mlflow.utils._is_kedro_project", return_value=True)
+    monkeypatch.chdir(tmp_path)
+
+    context = load_context(tmp_path)
+    mlflow_conf = get_mlflow_config(context)
+    mlflow.set_tracking_uri(mlflow_conf.mlflow_tracking_uri)
+
+    with mlflow.start_run():
+        existing_run_id = mlflow.active_run().info.run_id
+
+    dummy_catalog_with_run_id = DataCatalog(
+        {
+            "raw_data": MemoryDataSet(pd.DataFrame(data=[1], columns=["a"])),
+            "params:unused_param": MemoryDataSet("blah"),
+            "data": MemoryDataSet(),
+            "model": PickleDataSet((tmp_path / "model.csv").as_posix()),
+            "my_metrics": MlflowMetricsDataSet(run_id=existing_run_id),
+            "another_metrics": MlflowMetricsDataSet(
+                run_id=existing_run_id, prefix="foo"
+            ),
+        }
+    )
+
+    pipeline_hook = MlflowPipelineHook()
+
+    runner = SequentialRunner()
+    pipeline_hook.after_catalog_created(
+        catalog=dummy_catalog_with_run_id,
+        # `after_catalog_created` is not using any of arguments bellow,
+        # so we are setting them to empty values.
+        conf_catalog={},
+        conf_creds={},
+        feed_dict={},
+        save_version="",
+        load_versions="",
+        run_id=dummy_run_params["run_id"],
+    )
+    pipeline_hook.before_pipeline_run(
+        run_params=dummy_run_params,
+        pipeline=dummy_pipeline_ml,
+        catalog=dummy_catalog_with_run_id,
+    )
+    runner.run(dummy_pipeline_ml, dummy_catalog_with_run_id)
+
+    current_run_id = mlflow.active_run().info.run_id
+
+    pipeline_hook.after_pipeline_run(
+        run_params=dummy_run_params,
+        pipeline=dummy_pipeline_ml,
+        catalog=dummy_catalog_with_run_id,
+    )
+
+    mlflow_client = MlflowClient(mlflow_conf.mlflow_tracking_uri)
+    all_runs_id = set(
+        [run.run_id for run in mlflow_client.list_run_infos(experiment_id="0")]
+    )
+
+    # the metrics are supposed to have been logged inside existing_run_id
+    run_data = mlflow_client.get_run(existing_run_id).data
+
+    # Check if metrics datasets have prefix with its names.
+    # for metric
+    assert all_runs_id == {current_run_id, existing_run_id}
+    assert run_data.metrics["my_metrics.metric_key"] == 1.1
+    assert run_data.metrics["foo.metric_key"] == 1.1
+
+
+@pytest.mark.parametrize(
+    "model_signature,expected_signature",
+    (
+        [None, None],
+        ["auto", pytest.lazy_fixture("dummy_signature")],
+        [
+            pytest.lazy_fixture("dummy_signature"),
+            pytest.lazy_fixture("dummy_signature"),
+        ],
+    ),
+)
+def test_mlflow_pipeline_hook_with_pipeline_ml_signature(
+    mocker,
+    monkeypatch,
+    tmp_path,
+    config_dir,
+    env_from_dict,
+    dummy_pipeline,
+    dummy_catalog,
+    dummy_run_params,
+    dummy_mlflow_conf,
+    model_signature,
+    expected_signature,
+):
+    # config_with_base_mlflow_conf is a conftest fixture
+    mocker.patch("kedro_mlflow.utils._is_kedro_project", return_value=True)
+    monkeypatch.chdir(tmp_path)
+    pipeline_hook = MlflowPipelineHook()
+    runner = SequentialRunner()
+
+    pipeline_to_run = pipeline_ml_factory(
+        training=dummy_pipeline.only_nodes_with_tags("training"),
+        inference=dummy_pipeline.only_nodes_with_tags("inference"),
+        input_name="raw_data",
+        conda_env=env_from_dict,
+        model_name="model",
+        model_signature=model_signature,
+    )
+
+    pipeline_hook.after_catalog_created(
+        catalog=dummy_catalog,
+        # `after_catalog_created` is not using any of arguments bellow,
+        # so we are setting them to empty values.
+        conf_catalog={},
+        conf_creds={},
+        feed_dict={},
+        save_version="",
+        load_versions="",
+        run_id=dummy_run_params["run_id"],
+    )
+    pipeline_hook.before_pipeline_run(
+        run_params=dummy_run_params, pipeline=pipeline_to_run, catalog=dummy_catalog
+    )
+    runner.run(pipeline_to_run, dummy_catalog)
+    run_id = mlflow.active_run().info.run_id
+    pipeline_hook.after_pipeline_run(
+        run_params=dummy_run_params, pipeline=pipeline_to_run, catalog=dummy_catalog
+    )
+
+    # test : parameters should have been logged
+    trained_model = mlflow.pyfunc.load_model(f"runs:/{run_id}/model")
+    assert trained_model.metadata.signature == expected_signature
 
 
 def test_generate_kedro_commands():