+++ tests/io/metrics/test_mlflow_metrics_dataset.py
@@ -2,10 +2,11 @@ from typing import Dict, List, Optional, Union
 
 import mlflow
 import pytest
+from kedro.io import DataSetError
 from mlflow.tracking import MlflowClient
 from pytest_lazyfixture import lazy_fixture
 
-from kedro_mlflow.io import MlflowMetricsDataSet
+from kedro_mlflow.io.metrics import MlflowMetricsDataSet
 
 
 def assert_are_metrics_logged(
@@ -105,3 +106,66 @@ def test_mlflow_metrics_dataset_saved_and_logged(tmp_path, tracking_uri, data, p
     for k in catalog_metrics.keys():
         data_key = k.split(".")[-1] if prefix is not None else k
         assert data[data_key] == catalog_metrics[k]
+
+
+def test_mlflow_metrics_dataset_saved_without_run_id(tmp_path, tracking_uri, metrics3):
+    """Check if MlflowMetricsDataSet can be saved in catalog when filepath is given,
+    and if logged in mlflow.
+    """
+    prefix = "test_metric"
+
+    mlflow.set_tracking_uri(tracking_uri.as_uri())
+    mlflow_client = MlflowClient(tracking_uri=tracking_uri.as_uri())
+    mlflow_metrics_dataset = MlflowMetricsDataSet(prefix=prefix)
+
+    # a mlflow run_id is automatically created
+    mlflow_metrics_dataset.save(metrics3)
+    run_id = mlflow.active_run().info.run_id
+
+    assert_are_metrics_logged(metrics3, mlflow_client, run_id, prefix)
+
+
+def test_mlflow_metrics_dataset_exists(tmp_path, tracking_uri, metrics3):
+    """Check if MlflowMetricsDataSet is well identified as
+    existing if it has already been saved.
+    """
+    prefix = "test_metric"
+
+    mlflow.set_tracking_uri(tracking_uri.as_uri())
+    mlflow_metrics_dataset = MlflowMetricsDataSet(prefix=prefix)
+
+    # a mlflow run_id is automatically created
+    mlflow_metrics_dataset.save(metrics3)
+    assert mlflow_metrics_dataset.exists()
+
+
+def test_mlflow_metrics_dataset_does_not_exist(tmp_path, tracking_uri, metrics3):
+    """Check if MlflowMetricsDataSet is well identified as
+    not existingif it has never been saved.
+    """
+
+    mlflow.set_tracking_uri(tracking_uri.as_uri())
+    mlflow.start_run()  # starts a run toenable mlflow_metrics_dataset to know where to seacrh
+    run_id = mlflow.active_run().info.run_id
+    mlflow.end_run()
+    mlflow_metrics_dataset = MlflowMetricsDataSet(prefix="test_metric", run_id=run_id)
+    # a mlflow run_id is automatically created
+    assert not mlflow_metrics_dataset.exists()
+
+
+def test_mlflow_metrics_dataset_fails_with_invalid_metric(
+    tmp_path, tracking_uri, metrics3
+):
+    """Check if MlflowMetricsDataSet is well identified as
+    not existingif it has never been saved.
+    """
+
+    mlflow.set_tracking_uri(tracking_uri.as_uri())
+    mlflow_metrics_dataset = MlflowMetricsDataSet(prefix="test_metric")
+
+    with pytest.raises(
+        DataSetError, match="Unexpected metric value. Should be of type"
+    ):
+        mlflow_metrics_dataset.save(
+            {"metric1": 1}
+        )  # key: value is not valid, you must specify {key: {value, step}}